# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Server API 文档",
    "version": "0.0.1",
    "description": "床前明月光，疑是地上霜。举头望明月，低头思故乡。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/api"
    }
  ],
  "paths": {
...
    "/users/visitor-login": {
      "post": {
        "summary": "游客登录和注册",
        "description": "允许游客使用账号和密码登录或注册",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "account": {
                    "type": "string",
                    "description": "用户账号",
                    "required": true
                  },
                  "password": {
                    "type": "string",
                    "description": "用户密码",
                    "required": true
                  },
                  "nickname": {
                    "type": "string",
                    "description": "可选的用户昵称"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录/注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "account": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "profile": {
                          "type": "object",
                          "properties": {
                            "nickname": {
                              "type": "string",
                              "description": "用户昵称"
                            },
                            "avatar": {
                              "type": "string",
                              "description": "头像URL"
                            },
                            "phone": {
                              "type": "string",
                              "description": "电话号码"
                            },
                            "email": {
                              "type": "string",
                              "description": "电子邮箱"
                            },
                            "bio": {
                              "type": "string",
                              "description": "个人简介"
                            },
                            "gender": {
                              "type": "string",
                              "enum": ["male", "female", "other"],
                              "description": "性别"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "账号权限不适合"
          },
          "500": {
            "description": "登录失败"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "用户登录",
        "description": "非游客用户登录接口",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account", "password"],
                "properties": {
                  "account": {
                    "type": "string",
                    "description": "用户账号"
                  },
                  "password": {
                    "type": "string",
                    "description": "用户密码"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "account": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "profile": {
                          "type": "object",
                          "properties": {
                            "nickname": {
                              "type": "string",
                              "description": "用户昵称"
                            },
                            "avatar": {
                              "type": "string",
                              "description": "头像URL"
                            },
                            "phone": {
                              "type": "string",
                              "description": "电话号码"
                            },
                            "email": {
                              "type": "string",
                              "description": "电子邮箱"
                            },
                            "bio": {
                              "type": "string",
                              "description": "个人简介"
                            },
                            "gender": {
                              "type": "string",
                              "enum": ["male", "female", "other"],
                              "description": "性别"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "登录失败"
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "summary": "用户注册",
        "description": "非游客用户注册接口",
        "tags": ["Users"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account", "password"],
                "properties": {
                  "account": {
                    "type": "string",
                    "description": "用户账号"
                  },
                  "password": {
                    "type": "string",
                    "description": "用户密码"
                  },
                  "role": {
                    "type": "string",
                    "description": "用户角色（默认为member）",
                    "default": "member"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "注册失败"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "summary": "用户退出登录",
        "description": "注销当前用户的会话",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "退出登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "退出登录失败"
          }
        }
      }
    },
    "/users/search": {
      "get": {
        "summary": "搜索用户",
        "description": "搜索用户",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string",
              "default": 1
            },
            "required": true,
            "description": "用户昵称或者账号"
          }
        ],
        "responses": {
          "200": {
            "description": "搜索成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "account": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "profile": {
                          "type": "object",
                          "properties": {
                            "nickname": {
                              "type": "string",
                              "description": "用户昵称"
                            },
                            "avatar": {
                              "type": "string",
                              "description": "头像URL"
                            },
                            "phone": {
                              "type": "string",
                              "description": "电话号码"
                            },
                            "email": {
                              "type": "string",
                              "description": "电子邮箱"
                            },
                            "bio": {
                              "type": "string",
                              "description": "个人简介"
                            },
                            "gender": {
                              "type": "string",
                              "enum": ["male", "female", "other"],
                              "description": "性别"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "更新失败"
          }
        }
      }
    },
    "/users/profile": {
      "put": {
        "summary": "更新用户资料",
        "description": "更新当前登录用户的个人资料",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "description": "用户昵称"
                  },
                  "avatar": {
                    "type": "string",
                    "description": "头像URL"
                  },
                  "phone": {
                    "type": "string",
                    "description": "电话号码"
                  },
                  "email": {
                    "type": "string",
                    "description": "电子邮箱"
                  },
                  "bio": {
                    "type": "string",
                    "description": "个人简介"
                  },
                  "gender": {
                    "type": "string",
                    "enum": ["male", "female", "other"],
                    "description": "性别"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "资料更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "account": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "profile": {
                          "type": "object",
                          "properties": {
                            "nickname": {
                            "type": "string",
                            "description": "用户昵称"
                          },
                          "avatar": {
                            "type": "string",
                            "description": "头像URL"
                          },
                          "phone": {
                            "type": "string",
                            "description": "电话号码"
                          },
                          "email": {
                            "type": "string",
                            "description": "电子邮箱"
                          },
                          "bio": {
                            "type": "string",
                            "description": "个人简介"
                          },
                          "gender": {
                            "type": "string",
                            "enum": ["male", "female", "other"],
                            "description": "性别"
                          }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "更新失败"
          }
        }
      }
    },
    "/users/refresh-token": {
      "get": {
        "summary": "刷新认证token",
        "description": "使用refresh token获取新的access token",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "access_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "refresh token无效"
          },
          "500": {
            "description": "刷新失败"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}